feat: add feature version history in library detail

+++ Modules/Smartlegal/Http/Controllers/DocMandatoryController.php
@@ -55,6 +55,7 @@ class DocMandatoryController extends Controller
                 'f.intFileID', 'f.txtFilename', 'f.txtPath',
                 'e2.txtDepartmentName AS txtCostCenterName', 'e2.txtInitial AS txtCostCenterInitial'
             ])
+            ->orderBy('d.intDocID', 'DESC')
             ->get();
     
             $transformedData = $data->map(function ($row) {
@@ -297,7 +298,7 @@ class DocMandatoryController extends Controller
         if ($request->hasFile('txtFile')) {
             if($file->txtFilename != 'default.pdf') {
                 $destroy = public_path($file->txtPath);
-                unlink($destroy);
+                if (file_exists($destroy)) unlink($destroy);
             }
             $reqFile = $request->file('txtFile');
             $fileName = time() . '_' . $reqFile->getClientOriginalName();
